First commit of public branch of thesis source code.
This repo includes many *generated* files (for example
`./figure/*.{tex,pdf}`, as well as `thesis.pdf` itself)
as a courtesy in order to provide immediate access to these
files for anyone without requiring them to first recompile
the thesis.

If you plan to use this repo as a starting point for your own
document, you should probably remove those generated files from
version control (see .gitignore and search the web on how to
accomplish that) to avoid bloating your own git repo.
(As a general rule, any file that is *generated* as part of
your compile/build process should not be tracked by git).